backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Timur Berezin", email = "corporate.timur01@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
